# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tournament": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Get a list of tournaments",
                "operationId": "fd7d76c50ddd21da426fea7e201c3a8f",
                "parameters": [
                    {
                        "name": "filters",
                        "in": "query",
                        "description": "Array of filters",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "type": {
                                        "description": "Filter type",
                                        "type": "string"
                                    },
                                    "value": {
                                        "description": "Filter value",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by field",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Create a new tournament",
                "operationId": "15ceafb6ecc764d49e8c7af5c678f8ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Summer Tournament"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ],
                                        "example": "male"
                                    },
                                    "players": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "dni": {
                                                    "type": "integer",
                                                    "example": 12345678
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "enum": [
                                                        "male",
                                                        "female"
                                                    ],
                                                    "example": "male"
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "age": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "skill": {
                                                    "type": "number",
                                                    "example": 85.5
                                                },
                                                "strength": {
                                                    "type": "number",
                                                    "example": 75
                                                },
                                                "speed": {
                                                    "type": "number",
                                                    "example": 80
                                                },
                                                "latency": {
                                                    "type": "number",
                                                    "example": 50
                                                },
                                                "experience": {
                                                    "type": "number",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tournament created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/tournament/{id}/play": {
            "post": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "play a tournament",
                "operationId": "8548433bd0e22b3013efb315c38bad27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player Winner",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/tournament/{id}": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "show a tournament",
                "operationId": "c470f6a4383a45699f3a11b34f0e304a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournament",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Tournaments",
            "description": "Tournaments"
        }
    ]
}